moved some properties and methods of player in a parent classes
